+++ plugin/src/main/groovy/org/openbakery/appstore/AbstractAppstoreTask.groovy
@@ -13,16 +13,16 @@ class AbstractAppstoreTask extends AbstractDistributeTask {
 	def runAltool(String action) {
 		File ipa = getIpaBundle()
 
-		if (project.appstore.username == null) {
-			throw new IllegalArgumentException("Appstore username is missing. Parameter: appstore.username")
+		if (project.appstore.apiKey == null) {
+			throw new IllegalArgumentException("Appstore apiKey is missing. Parameter: appstore.apiKey")
 		}
 
-		if (project.appstore.password == null) {
-			throw new IllegalArgumentException("Appstore password is missing. Parameter: appstore.password")
+		if (project.appstore.apiIssuer == null) {
+			throw new IllegalArgumentException("Appstore apiIssuer is missing. Parameter: appstore.apiIssuer")
 		}
 
 		StyledTextOutput output = getServices().get(StyledTextOutputFactory.class).create(AbstractAppstoreTask.class)
 
-		commandRunner.run([xcode.getAltool(), action, "--username", project.appstore.username, "--password",  project.appstore.password, "--file", ipa.getAbsolutePath()], new ConsoleOutputAppender(output))
+		commandRunner.run([xcode.getAltool(), action, "--apiKey", project.appstore.apiKey, "--apiIssuer",  project.appstore.apiIssuer, "--file", ipa.getAbsolutePath()], new ConsoleOutputAppender(output))
 	}
 }