+++ plugin/src/main/groovy/org/openbakery/appstore/AppstorePluginExtension.groovy
@@ -4,8 +4,8 @@ import org.gradle.api.Project
 
 class AppstorePluginExtension {
 
-	def String username = null
-	def String password = null
+	def String apiKey = null
+	def String apiIssuer = null
 	def String ascProvider = null
 
 	public AppstorePluginExtension(Project project) {